requirements.txt
blinker==1.9.0
click==8.3.1
dotenv==0.9.9
Flask==3.1.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
